+++ chrontext/src/rewriting/graph_patterns/left_join_pattern.rs
@@ -2,6 +2,7 @@ use super::StaticQueryRewriter;
 use crate::change_types::ChangeType;
 use crate::query_context::{Context, PathEntry};
 use crate::rewriting::graph_patterns::GPReturn;
+use crate::rewriting::subqueries::{SubQuery, SubQueryInContext};
 use spargebra::algebra::{Expression, GraphPattern};
 
 impl StaticQueryRewriter {
@@ -29,39 +30,41 @@ impl StaticQueryRewriter {
                 || right_rewrite.is_subquery
                 || !expression_rewrite.pushups.is_empty()
             {
+                let left_subquery_context;
                 if !left_rewrite.is_subquery {
-                    self.create_add_subquery(
-                        left_rewrite,
-                        &left_context,
-                        PathEntry::LeftJoinLeftSide,
-                    );
+                    self.create_add_subquery(left_rewrite, &left_context);
+                    left_subquery_context = left_context.clone();
+                } else {
+                    left_subquery_context = left_rewrite.subquery_context.unwrap().clone();
                 }
+                let right_subquery_context;
                 if !right_rewrite.is_subquery {
                     self.create_add_subquery(
                         right_rewrite,
                         &right_context,
-                        PathEntry::LeftJoinRightSide,
                     );
+                    right_subquery_context = right_context.clone();
+                } else {
+                    right_subquery_context = right_rewrite.subquery_context.unwrap().clone();
                 }
-
-                let mut subquery_vec = vec![
-                    (PathEntry::LeftJoinLeftSide, left_context),
-                    (PathEntry::LeftJoinRightSide, right_context),
-                ];
+                let mut expression_contexts = vec![];
                 for (gp, ctx) in expression_rewrite
                     .pushups
                     .iter()
                     .zip(expression_rewrite.pushup_contexts.iter())
                 {
                     //TODO: Fix the context so these things are recoverable..
-                    self.create_add_subquery(
-                        gp.clone(),
-                        &expression_context,
-                        PathEntry::LeftJoinExpression,
-                    );
-                    subquery_vec.push((PathEntry::LeftJoinExpression, expression_context.clone()));
+                    self.create_add_subquery(gp.clone(), ctx);
+                    expression_contexts.push(ctx.clone());
                 }
-                self.subquery_ntuples.push(subquery_vec);
+                self.subqueries_in_context.push(SubQueryInContext::new(
+                    context.clone(),
+                    SubQuery::LeftJoin(
+                        left_subquery_context,
+                        right_subquery_context,
+                        expression_contexts,
+                    ),
+                ));
                 let mut ret = GPReturn::subquery(context.clone());
                 return ret;
             } else {
@@ -82,27 +85,28 @@ impl StaticQueryRewriter {
                 return left_rewrite;
             }
         } else {
-            if left_rewrite.is_subquery
-                || right_rewrite.is_subquery
-            {
+            if left_rewrite.is_subquery || right_rewrite.is_subquery {
+                let left_subquery_context;
                 if !left_rewrite.is_subquery {
-                    self.create_add_subquery(
-                        left_rewrite,
-                        &left_context,
-                        PathEntry::LeftJoinLeftSide,
-                    );
+                    self.create_add_subquery(left_rewrite, &left_context);
+                    left_subquery_context = left_context.clone();
+                } else {
+                    left_subquery_context = left_rewrite.subquery_context.unwrap().clone();
                 }
+                let right_subquery_context;
                 if !right_rewrite.is_subquery {
                     self.create_add_subquery(
                         right_rewrite,
                         &right_context,
-                        PathEntry::LeftJoinRightSide,
                     );
+                    right_subquery_context = right_context.clone();
+                } else {
+                    right_subquery_context = right_rewrite.subquery_context.unwrap().clone();
                 }
-                self.subquery_ntuples.push(vec![
-                    (PathEntry::LeftJoinLeftSide, left_context),
-                    (PathEntry::LeftJoinRightSide, right_context),
-                ]);
+                self.subqueries_in_context.push(SubQueryInContext::new(
+                    context.clone(),
+                    SubQuery::LeftJoin(left_subquery_context, right_subquery_context, vec![]),
+                ));
 
                 let mut ret = GPReturn::subquery(context.clone());
                 return ret;
